--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "prompt-sync": "^4.2.0"
-  }
-}
