--- step_on_step/app/schedule-app/src/services/express.js
@@ -1,38 +1,0 @@
-const express = require("express");
-
-const server = express();
-
-server.use(express.json());
-
-const contacts = ["Varela", "Robert", "Lukita"];
-
-server.get("/contacts", (req, res) => {
-  const { index } = req.params;
-
-  return res.json(contacts[index]);
-});
-
-server.get("/contacts", (req, res) => {
-  return res.json(contacts);
-});
-
-server.post("/contacts", (req, res) => {
-  const { name } = req.body;
-
-  return res.json(contacts.append(name));
-});
-
-server.put("/contacts", (req, res) => {
-  const { index } = req.body;
-  const { name } = req.body;
-
-  return res.json((contacts[index] = name));
-});
-
-server.delete(`/contacts/${id}`, (req, res) => {
-  const { index } = req.body;
-
-  return res.json(contacts.splice(index, 1));
-});
-
-server.listen(5000);
